requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
python-dotenv
google-api-python-client
psycopg2-binary
gunicorn
